requirements.txt
Flask==3.0.3
pandas==1.5.3
PyYAML==6.0
numpy==1.23.5
requests==2.28.1
gunicorn==20.1.0